Refactor: some imports fixed
