+++ src/main/java/com/optivision/webapp/sale/entity/SaleItemEntity.java
@@ -1,4 +1,4 @@
-package com.optivision.webapp.venta.entity;
+package com.optivision.webapp.sale.entity;
 
 import lombok.AllArgsConstructor;
 import lombok.Builder;
